Add a per-entry custom user data field.
This is useful to allow leaders to pass context between
raft_recv_entry() and a subsequent execution of the entry (after it was
committed) as an applylog callback.

This allows an optimized use pattern when entries received locally on
the leader can be used without going through an unnecessary
deserialization.