Better raftize-all-commands support.
As the latest unstable Redis now supports the command filtering API,
this feature is no longer an #ifdef.

To maintain compatibility we make sure the API is implemented and fail
gracefully if not.

Also, `RAFT.CONFIG SET` now supports enabling and disabling this feature
at runtime and not just on startup.