Fix raftize re-entrancy issues (#13, #14).
Redis Module API makes sure calls from the module back to Redis don't
get re-intercepted by the filtering hook. However, this guarantee is
valid for calls made on a normal context, but not on a thread safe
context.

In many cases Raft log entries are applied in a different context, which
causes the filter to re-raftize them. This leads with an endless loop
that bloats the log and does causes log entries to never apply to Redis.