Fix issues around node startup and log loading.
During log loading, historic nodes could be resurrected and communicated
with; this is bad in particular if such nodes exist, or if a node ends
up sending messages to itself because the old node shared the same
endpoint.

Commit index not always properly initialized when the log indicates a
single voting node exists.