Add response timeout mechanism.
Fixes issues with sporadic loss of quorum due to load, in particular
when follower proxy is used.

Proxied commands from followers introduce head-of-line blocking on the
connection, which may delay or completely block Raft messages (such as
RequestVote, or AppendEntries when followers are promoted).

This introduces a timeout mechanism that tracks queued commands:
1. Proxy response timeout refers to commands that have been proxied as
RAFT.ENTRY from folllowers to leaders.
2. Raft response timeout refers to Raft commands.

When a response timeout is reached, the connection is considered blocked
and will be dropped and re-established.

This fix also changes the default Raft request and election timeouts to
allow for greater number of retries before calling an election.