Fix config loading from snapshots on startup.
This fixes an issue with loading of node configuration from snapshot to
the Raft library, when RedisRaft re-starts and loads from snapshot.

Due to missing initalization, the Raft library remained in a state where
the loaded configuration was *working*, but an attempt to re-create a
"2nd generation" snapshot from it resulted with missing nodes.

This may have been obscured by #33, as well as by the fact that
additional node membership operations, if done before a snapshot is
taken, would correct this state.

Seems to be the root cause for #44.