Fix occasional crashes after snapshots.
When a snapshot is taken the log is rewritten to begin right after the
last applied entry. The process of switching the log got its index
truncated, making all existing entries inaccessible - an attempt to
access them would assert.

Note: truncation *is* generally desired because the log and the index
cannot be atomically updated in the filesystem. For this reason, we
generally truncate and re-build the index when opening the log on
process start. In this case however the index should have been reused
(or at least re-built, although there's really no reason to do that).

This seems to be the root cause for #42 and possibly #43 as well.