Move log rewrite from child process.
Before this, log rewrite was handled by the child process immediately
after a snapshot (RDB file) was created, followed by appending any newly
received entries that did not exist at the time of forking.

This mechanism does not address removal of entires from the log during
this process, so it is now reverted and the entire log is rewritten by
the parent after the child completes.

In the future we can revisit this by creating a more complex mechanism
to track "low water mark". It is not clear how effective this will be,
as the portion of the log following the snapshot should anyway be fairly
small in most cases.