Fix an off-by-one issue with RaftLogImpl.
The API specifies the index passed on reset is the one that's going to
be used for the *next* entry.

As a result, the truncated log was created with an off-by-one header in
handleLoadSnapshot(). However, RaftLogCreate() was immediately called
with the proper index, masking the problem.

If the process was restarted in-between, the corrupt header would be
read causing the log index to be ahead of the real log by a single
entry. Next, raft_periodic() would be called, attempt to read the last
log entry, fail and trigger an assert.

This fixes the last (known?) case that manifested as #42.