Initial shardgroup update support.
For every shardgroup, we maintain a Connection object that is used to
communicate with the leader node on the remote cluster.

The connection uses the new connections infrastructure to automatically
handle reconnects. All known nodes are iterated, and -MOVED responses
are considered as well when attempting to locate the leader.

A new cluster periodic call function is used to re-trigger updates on
live connections.

Also, add RAFT.SHARDGROUP LINK support.

This simplifies the process of joining a remote cluster shardgroup by
managing the connection and shardgroup configuration retrieval
automatically.