Update tests.
Previously, most tests assumed a default of raftize-all-commands
disabled. The RedisRaft.raft_exec() method was extensively used to wrap
a command with 'RAFT' and deliver it directly to the module.

The exception were some tests that were specifically written to test
raftize-all-commands behavior, or more modern tests that were already
created with the anticipation of raftize-all-commands going away at some
point.

In a nutshell, changes follow these principles:
* Instead of calling RedisRaft.raft_exec() we use redis-py's native
  command.
* Cluster.raft_exec() was renamed to Cluster.execute() to avoid
  confusion. It also does not explicitly call the 'RAFT' command, but
  does handle retries, MOVED redirects, etc. so it can still be useful.
* In some cases, tests did issue direct Redis commands in order to
  directly inspect the state of the dataset on a specific replica
  (without going through consensus reads, etc.). This is now done by
  calling RedisRaft.raft_debug_exec() which uses the RAFT.DEBUG EXEC
  mechanism to bypass command interception.