Move snapshot out of keyspace.
Before this commit, snapshot information was persisted by creating a
module data type and having a fake key in the keyspace. The RDB save and
load callbacks would then be triggered to store or load the information.

With new versions of the module API that support RDB AUX saving and
loading, it is no longer necessary to have a fake key.