Multiple Virtraft changes (#8)
* fix virtraft bug in _check_log_matching

didn't validate against snapshoted node correctly.  the previous code just didn't do what it was expected to do

* improve virtraft logging

* fix entry_pop in virtraft

previously, entry_pop never ran during our virtraft tests, probably because without removing leader, we never have a situation where this is neccessary

i.e. the only case virtraft can pop entries, is when we have a leader lose leadership and we have an entry pending and not committed.  so when another node becomes leader, old leader's entries get popped.

however, that has nothing to do with the "id" of the node that its effecting so trying to contorl it through that, makes no sense.

* extra comment for fix reminder

* ensure that we start with a term of 1 in virtraft, not 0.

snapshot code fails if we have a snapshot term of 0 (explicitly returns an error if the term stored in snapshot is 0).

Co-authored-by: Yossi Gottlieb <yossigo@gmail.com>