add a read_queue test to virtraft (#41)
* add a read_queue test to virtraft

every iteration we push a read_queue request and the handler and an we pass to it set a variable on calling.  we can use this to make sure that the read_queue doesn't get too far from the iteration.  i.e. we pass the leader's msg_id and can check to ensure that leader's msg_id doesn't get too far from the msg_id (variable) we see in read_queue test

this is analog to the current log applying deadlock test.

* implements msg_id checking in virtraft for verification of read_queue requests

When we pop an entry off the read_queue with the can_read flag, we verify across all nodes that the majority of nodes have accepted from the leader a msg_id past what this read_queue entry needs.

the problem with this is, that until now, msg_id was private to each server instance and has no relevance to the followre nodes except to include it back in response. 

we change that to have the followers store the max msg_id they've seen from their current leader

using that, in the read_queue_handler we can verify that the leader's voting nodes have a quorum past the msg_id variable that the handler will return to ensure that this read_queue handler call is correct.