Enable core dumps on Jepsen runs.
This configures the kernel to generate local core dumps, instead of
piping them through apport which is the default.

Debian stretch seems to have an unlimited core file size by default, so
crashes in Redis should generate a core dump.

And the latest version of the RedisRaft jepsen is now set to locate core
files and pack them along wiht log files.