Add ignore-commands config option to Redis Raft (#188)
This adds the ignore-commands configuraiton option on the command line that takes a comma delimited list of command that Redis Raft should ignore.

ex: "ignore-commands command1,command2"

included is an integration test that compares an ignored command issued against an unitialized cluster vs a non ignored command.  An ignored command will fail in the normal redis manner, while the a non ignored command will attempt to be applied to the raft log and will fail because the cluster hasn't been initialized yet.