remove shardgroup-id config and just use generated dbid
the only negative is that the shardgroup created for the local cluster will not have sg->id filled in, as created on redis startup, but dbid is only initialized on cluster init.

this isn't a major problem, as it will always be array item 0 and we treat that specially anyways.