WIP: Can we intercept Lua commands?
If we use REDISMODULE_CMDFILTER_NOSELF we can't, because we're the ones
to initiate the RM_Call() for EVAL/EVALSHA.

We can give up this flag but then we need to be more careful with
re-entrancy ourselves.

Also, need to be able to determine that we're in a Lua context. Two
options:
* Track flows where we initiate RM_Call() for an EVAL/EVALSHA and set a
  flag. Easy but probably not 100% fool proof (e.g. RM_Call -> third
  party module -> Lua).
* Extend the filter context to tell us this. It's deliberately not a
  full RedisModuleCtx to avoid abuse, but adding a few more flags should
  be harmless.