further improvement on ShardGroup ownership semantics
previously a ShardGroup was created, then passed to ShardGroupDeserialize()
this made ownership semantics ugly.

now, ShardGroupDeserialize() does the ShardGroupCreation, and if there's an error, it frees it, otherwise it returns it, passing full ownership to the caller.

similarly, as before, the caller then passses it to another consuming function, which it has transfered ownership to, never to get it back.