add internode auth to redisraft
currently, there is no authentication that is done between RedisRaft clusters

this adds the ability to define which acl user and what password should be used, as has redis authenticate them

the default user (if not specified) is default, and the default password is empty.

when used, on must ensure that redis itself is configured to accept such user/password authenticaiton.
If using the default user of default, one can just pass a --requirepass to redis, if one is changing the user, one must setup the acl user correctly in redis conf/acl files