remove my previous serialization changes + move all logic to apply
this removes all the shard / asking checks from the append stage on leader and pushes them all into the apply stage on every node.

negative is we no longer short cut cross slot functionality (side effect of moving everything to apply, its arguable that we can still check this at append time as not dependent on raft FSM).

we also cannot redirect to remove cluster at append time, as we only redirect to leader at append time.  its possible that this could be optimized as well if needed.