check command(s)if they are deny-oom and deny if in an oom / unknown situation.
currently populates a deny-oom dict on demand (mostly to handle modules being loaded after redisraft that insert their own commands).  Can populate at init time only or a combination of the two (i.e. init what we know, and if we see a command we don't know, init it on demand)