refactor handleRedisCommand so asking and proxy can work
split handleRedisCommand into 2 functions

handleRedisCommand and handleRedisCommandAppend

handleRedisCommand can intercept and reinterpert and modify the command, while handleRedisCommand is just the process to accept/reject the appending of the entry.

this way, proxied entries (via raft.entry) will just call handleRedisCommandAppend and will use the command exactly as they recieved it from the node the client is connected to (i.e. how that node would have appended it if it was leader)