Use SEEK_SET to set position of the snapshot file (#510)
Use SEEK_SET to set file position while storing snapshot chunks.

In raftStoreSnapshotChunk(), offset indicates position relative
to the head of the file.

Currently, we use SEEK_CUR which sets the position relative to the
current position. As we reopen file each time, current position is zero
and it works correctly but better to use SEEK_SET which sets position
relative to the head of the file. (Just to prevent future bugs).