Do not print TLS error string if a request fails (#504)
We should handle TLS connections similar to TCP. If a TCP connection
gets disconnected, we don't try to print the details. It should be the
same for TLS connections. 

The only moment we want to print TLS specific error message is when we
fail to connect. We already print the error string from OpenSSL
(hiredis gives it to us) if there is an error while initiating the
connection.