Prevent deadlock while killing servers in tests (#524)
I've seen some tests that were hung for hours. After the investigation, 
one suspect might be: 

https://docs.python.org/3/library/subprocess.html#subprocess.Popen.wait

```
Note This will deadlock when using stdout=PIPE or stderr=PIPE
and the child process generates enough output to a pipe such that 
it blocks waiting for the OS pipe buffer to accept more data. 
Use Popen.communicate()when using pipes to avoid that.
```
I'm not sure if this was the reason for hung tests but documentation 
above suggests it is possible. 

- Changed `subprocess.wait()` calls to `subprocess.communicate()` 
- Added some more timeout / exception changes.