[RR-116] Do not load snapshot inside the command callback (#519)
RedisRaft crashes while loading a big snapshot.

RedisRaft calls `rdbLoad()` function from Redis to load the received
snapshot file (RDB). This function is called inside a command callback.
`rdbLoad()` goes to networking and process events if loading the 
snapshot takes a while. If the current client that triggered command
callback disconnects, client struct can be freed before it returns
from `rdbLoad()` and from the callback. After that, a use-after-free
bug occurs and leads to crash.

Previously, there was a similar bug. Redis was trying to process
another message from the same client:
https://github.com/RedisLabs/redisraft/pull/450

Until this issue is fixed, as a workaround, we can delay loading the
snapshot and let the command callback return. Inside the timer callback,
we can load the snapshot. To do that, we need to set state to 
`REDIS_RAFT_LOADING`, so, other commands should be rejected until we
load the snapshot.