Delete RAFT.DEBUG SENDSNAPSHOT (#540)
Currently, RAFT.DEBUG SENDSNAPSHOT is unused and it is leftover.
If a node is not lagging behind, it won't load the snapshot. As it has all
the entries, it will realize that snapshot is not necessary. So, this
command does not work.

In tests, we make target node lag behind (by killing it for a while) to
trigger snapshot delivery or we add a new node instead.