[RR-122] add crc32 value to every entry we write to the log (#539)
We create a "chain" of log entry "blocks", where each block in the chain is protected by a running crc32.  i.e. crc32 value of the block is the crc32 for all entries up to and including that block.  we start with the log header block and continue with all the log entries.

we implement tests that validate what happens if we corrupt the header and entries.
we implement tests to validate that we handle popping (deletion) of entries correctly as well as that we rebuild the running CRC correctly on restart by restarting multiple times.