[RR-184] Force close clients that exist while clearing sessions (#567)
In order to prevent clients with sessions from persisting if leadership transfers away and then returns, we force close them when clearing the session.

If we did not do this, a client that sent a WATCH could believe that its WATCH is still valid.  terminating the client connection will force the client
to reconnect.

To not require all clients to do this, we only close those that are in the middle of a session.