Trace level logs in integration tests (#585)
First commit fixes a bug. If raft.trace config is given on startup and
then we initialize the cluster, we don't pass log config to the raft library
after the initialization. This commit changes it and passes log config to
the raft library.
Second commit adds an option to pass raft.trace config to the pytest.
This is useful when we want detailed logs from a test.