[RR-174] Fix test_maxmemory_with_cluster_ops test failure (#586)
There were two issues with this test:

Leader adds nodes as non-voting first. Then, when leader
decides that it can promote a node to voting, it submits an entry to
make it voting. While that entry is in the log waiting to be
committed, we try to remove that node inside the test case. Libraft
rejects second configuration change entry as we don't allow multiple
configuration change entries at the same time.
To solve the problem, inside the test, we can wait until all nodes
become voting before trying to remove a node.

Second problem is when redisraft rejects removal operation, we
don't propagate the exception due to a bug. While checking for
something else, we ignore the exception by mistake. Added a change to
avoid that.