[RR-225] add "CLIENT UNBLOCK" and timeout blocking when disconnect
this add support for the CLIENT UNBLOCK command, so that we can send a timeout log entry when the user issues it, and will reply to the user correctly if we were able to unblock or not

in addition, it provides a best effort attempt at timing out entries when the blocked client disconnects.  However, it is possible that the blocked entry will execute before the timeout entry is applied, as is expected.