[RR-261] Free applied entries from the entry cache (#623)
Currently, we delete entries from the entry cache if memory usage goes
over configured size. This is happening inside periodic().

We may even delete entries that are waiting to be applied. When that 
happens, if an entry has a RaftReq entry associated with it, we'll 
reply with -TIMEOUT response when we remove the entry from the cache
and destroy it.

This is a bad user experience. Even though we'll apply the entry in a
moment, we return -TIMEOUT reply.

This PR changes this behavior, adds a check to prevent deleting entries
that are waiting to be applied.